fix:socket method divide
